refactor: refactored the managed cluster and managed control plane
The AWSManagedCluster has had most of its fields moved to the
AWSManagedControlPLane so that the reconciliiation logic is
cleaner. This follows the pattern used by CAPZ for the AKS
implementation.

The controller for AWSManagedCLuster just reports back the
control plane endpoint, ready and failure domains to CAPI
from the managed control plane.